initial commit of account_db functionality to branch
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/apidb/EbrcWebsiteCommon/branches/account_db@81734 735e2a04-f8fc-0310-8a1b-f2942603c481

+++ Model/src/main/java/org/eupathdb/common/controller/action/ContactUsAction.java
@@ -118,7 +118,7 @@ public class ContactUsAction extends WdkAction {
     DataHandler[] attachments = attachmentList.toArray(new DataHandler[0]);
 
     String supportEmail = wdkModel.getModelConfig().getSupportEmail();
-    String uid = Integer.toString(user.getUserId());
+    String uid = Long.toString(user.getUserId());
     // TODO: getVersion() reads from releaseVersion in model.xml which is not maintained
     String version = wdkModelBean.getBuild();
     String website = wdkModelBean.getDisplayName();
